requirements.txt
accelerate==0.21.0
addict==2.4.0
albumentations==1.4.0
anyio==3.7.1
beautifulsoup4==4.12.3
blendmodes==2024.1
caffe2==0.8.1
cityscapesscripts==2.2.2
clean_fid==0.1.35
cog==0.9.4
cv_bridge==1.13.0.post0
Cython==3.0.8
decord==0.6.0
dlib==19.24.2
dominate==2.9.1
dynamicprompts==0.30.4
einops==0.4.1
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.110.0
fire==0.5.0
fsspec==2024.2.0
ftfy==6.1.3
fvcore==0.1.5.post20221221
gdown==5.1.0
gfpgan==1.3.8
GitPython==3.1.32
GitPython==3.1.42
hydra-core==1.3.2
imutils==0.5.4
imwatermark==0.0.2
inflection==0.5.1
intel_extension_for_pytorch==2.2.0
iopath==0.1.9
jsonmerge==1.8.0
kornia==0.6.7
langchain==0.1.9
lark==1.1.2
lmdb==1.4.1
matplotlib==3.8.3
mc==0.0
mediapipe==0.10.10
mmdet==3.3.0
mmpose==1.3.1
natsort==8.4.0
numpy==1.23.5
omegaconf==2.2.3
onnx==1.15.0
onnxruntime==1.17.1
open_clip_torch==2.20.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
openvino==2023.3.0
packaging==23.2
pandas==2.2.1
parrots==1.0.3
piexif==1.1.3
Pillow==9.5.0
Pillow==10.2.0
pip==24.0
prettytable==3.10.0
psutil==5.9.5
pudb==2024.1
pycocoevalcap==1.2
pydantic==1.10.14
Pygments==2.17.2
pyTelegramBotAPI==4.16.1
pytest==8.0.2
python-dotenv==1.0.1
pytorch_fid==0.3.0
pytorch_lightning==1.9.4
pytz==2024.1
PyYAML==6.0.1
PyYAML==6.0.1
realesrgan==0.3.0
regex==2023.12.25
reportlab==4.1.0
Requests==2.31.0
resize_right==0.0.2
rich==13.7.1
ruamel_yaml==0.18.6
safetensors==0.3.1
scipy==1.12.0
sdata==0.19.1
selenium==4.18.1
Send2Trash==1.8.2
setuptools==59.6.0
Shapely==2.0.3
skimage==0.0
starlette==0.37.1
std_msgs==0.0.1
streamlit_drawable_canvas==0.9.3
svglib==1.5.1
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
termcolor==2.4.0
tifffile==2024.2.12
timm==0.9.2
tomesd==0.1.3
tomli==2.0.1
torch==2.0.1
torch_tb_profiler==0.4.3
torchdata==0.5.1
torchdiffeq==0.2.3
torchsde==0.2.6
torchvision==0.15.2
tqdm==4.66.2
transformers==4.30.2
turbojpeg==0.0.2
uvicorn==0.27.1
webdataset==0.2.86
webdriver_manager==4.0.1
webuiapi==0.9.8
xformers==0.0.24
yapf==0.40.2
